+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Themes: New default theme has been added, moreover users can choose between several themes and even create their own.
 - App: Adding new icons.
+- PWA: Adding pwa functionality. User can install the app on their devices (add to home screen).
 - Swagger: Adding Swagger documentation for the backends REST Api. Only for developer mode.
 - Developing: New '.env' variable added. Set 'DEV_USER' to 'true' and you can login with default user for developing purpose.
 - Dependencies: Bumping packages to newer versions.